Move MultipartFormData into separate file
